Update prime lib

+++ prime/prime.h
@@ -10,7 +10,7 @@
 
 unsigned *gen_prime_r (unsigned min, unsigned max, unsigned *p_size);
 long unsigned *gen_prime_ll_r (long unsigned min, long unsigned max, long unsigned *p_size);
-bool is_prime_r(long unsigned x, long unsigned arr[], long unsigned p_size);
+bool is_prime_r(unsigned x, unsigned arr[], unsigned p_size);
 bool is_prime_ll_r(long unsigned x, long unsigned arr[], long unsigned p_size);
 bool is_prime(long unsigned int n);
